Use a cache for index entries just written to disk
When confirms are used we flush the entire buffer to disk.
Because we are reading from the write_buffer when we can,
and that buffer becomes empty, we end up reading from disk
immediately after writing in some cases.

The cache helps avoid some of those reads from disk by keeping
the entries or acks in memory a little longer than before when
confirms are used.